Covid 19 Prediction
The Python script that I provided uses linear regression to predict the number of COVID-19 cases for the next 90 days. Linear regression is a simple machine learning algorithm that is used to model the relationship between a dependent variable (in this case, the number of COVID-19 cases) and one or more independent variables (in this case, the day number).

The script begins by importing the necessary libraries: pandas, numpy, and scikit-learn. Pandas is a library for data manipulation and analysis, numpy is a library for numerical computation, and scikit-learn is a library for machine learning.

Next, the script loads the COVID-19 cases data into a pandas dataframe using the read_csv function. It then splits the data into features (X) and target (y). The features are the day number, and the target is the number of cases.

The script then uses the train_test_split function from scikit-learn to split the data into training and testing sets. The training set is used to train the model, and the testing set is used to evaluate the model's performance.

After splitting the data, the script trains a linear regression model on the training data using the fit function. The model is then used to make predictions for the next 90 days using the predict function. Finally, the script prints the predictions to the console.

Keep in mind that this is just a sample script and the actual predictions will depend on the quality and completeness of the data that you use. It is always a good idea to validate the predictions using a separate test dataset and to fine-tune the model as needed.